clipboard-apis/preventing_default_action_paste_on_inputtext.html

<!DOCTYPE html>
<html><head>
    <meta charset="UTF-8">
    <title>Clipboard API - preventing default action   - paste on 'input_text'</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/_lib.js"></script>
    <script>
/** preventing default action   */
/* Events: paste */
/* Test HTML: <form><input id="input_text" autofocus onfocus="this.select()" value="original text"></form>*/
/* Targets: document 'input_text' */
/* paste data: "pasted text" */
function clipboard_api_test(e, test_obj){
    e.preventDefault();
    var theTarget=e.target;
    test_obj.step_timeout(function(){
      result(function(){assert_equals(theTarget.value, 'original text');});
    }, 10 );
}

    </script>
</head>
<body>

    <p>FAILED (This TC requires JavaScript enabled)</p>
    <form><input id="input_text" autofocus onfocus="this.select()" value="original text"></form>
    <div id="log"></div>
    <script>
    (function(){
        var t = async_test()
        var dataToPaste="pasted text" ;
        var extPassCond=null;
        var eventTarget='input_text';
        var eventType='paste';
        setup(function(){
                setupTest(t, eventTarget, eventType, dataToPaste, extPassCond );
            });
    })();
    </script>
</body></html>
